+++ _unittests/ut_mlmodel/test_sklearn_kmeans_constraint.py
@@ -206,9 +206,9 @@ class TestSklearnConstraintKMeans(ExtTestCase):
         try:
             pipe.fit(X_train, y_train)
         except AttributeError as e:
-                if compare_module_version(sklver, "0.24") < 0:
-                    return
-                raise e
+            if compare_module_version(sklver, "0.24") < 0:
+                return
+            raise e
         pred = pipe.predict(X_test)
         score = accuracy_score(y_test, pred)
         self.assertGreater(score, 0.8)